+++ lib/python/qmk/cli/info.py
@@ -10,7 +10,8 @@ from milc import cli
 from qmk.json_encoders import InfoJSONEncoder
 from qmk.constants import COL_LETTERS, ROW_LETTERS
 from qmk.decorators import automagic_keyboard, automagic_keymap
-from qmk.keyboard import keyboard_completer, keyboard_folder, render_layouts, render_layout, rules_mk
+from qmk.keyboard import keyboard_completer, keyboard_folder, rules_mk
+from qmk.render import render_layouts, render_layout
 from qmk.info import info_json, keymap_json
 from qmk.keymap import locate_keymap
 from qmk.path import is_keyboard
